backend/requirements.txt
-i https://pypi.org/simple
aiohttp==3.9.3; python_version >= '3.8'
aiosignal==1.3.1; python_version >= '3.7'
asgiref==3.7.2; python_version >= '3.7'
async-timeout==4.0.3; python_version >= '3.7'
attrs==23.2.0; python_version >= '3.7'
backports.zoneinfo==0.2.1; python_version >= '3.6'
base58==2.1.1; python_version >= '3.5'
bitarray==2.9.2
certifi==2023.11.17; python_version >= '3.6'
charset-normalizer==3.3.2; python_full_version >= '3.7.0'
cytoolz==0.12.3; python_version >= '3.7'
django==4.2.9; python_version >= '3.8'
djangorestframework==3.14.0; python_version >= '3.6'
eth-abi==5.0.0; python_version >= '3.8' and python_version < '4'
eth-account==0.10.0; python_version >= '3.7' and python_version < '4'
eth-hash[pycryptodome]==0.6.0; python_version >= '3.8' and python_version < '4'
eth-keyfile==0.7.0; python_version >= '3.8' and python_version < '4'
eth-keys==0.5.0; python_version >= '3.8' and python_version < '4'
eth-rlp==1.0.1; python_version >= '3.8' and python_version < '4'
eth-typing==4.0.0; python_version >= '3.8' and python_version < '4'
eth-utils==3.0.0; python_version >= '3.8' and python_version < '4'
frozenlist==1.4.1; python_version >= '3.8'
hexbytes==0.3.1; python_version >= '3.7' and python_version < '4'
idna==3.6; python_version >= '3.5'
importlib-resources==6.1.1; python_version >= '3.8'
ipfshttpclient==0.7.0; python_full_version >= '3.5.4' and python_full_version not in '3.6.0, 3.6.1, 3.7.0, 3.7.1'
jsonschema==4.21.1; python_version >= '3.8'
jsonschema-specifications==2023.12.1; python_version >= '3.8'
lru-dict==1.2.0
multiaddr==0.0.9; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
multidict==6.0.4; python_version >= '3.7'
netaddr==0.10.1
parsimonious==0.9.0
pkgutil-resolve-name==1.3.10; python_version >= '3.6'
protobuf==4.25.2; python_version >= '3.8'
pycryptodome==3.20.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
pytz==2023.4
pyunormalize==15.1.0; python_version >= '3.6'
referencing==0.33.0; python_version >= '3.8'
regex==2023.12.25; python_version >= '3.7'
requests==2.31.0; python_version >= '3.7'
rlp==4.0.0; python_version >= '3.8' and python_version < '4'
rpds-py==0.17.1; python_version >= '3.8'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sqlparse==0.4.4; python_version >= '3.5'
toolz==0.12.1; python_version >= '3.7'
typing-extensions==4.9.0; python_version >= '3.8'
urllib3==2.2.0; python_version >= '3.8'
varint==1.0.2
web3==6.15.0; python_full_version >= '3.7.2'
websockets==12.0; python_version >= '3.8'
yarl==1.9.4; python_version >= '3.7'
zipp==3.17.0; python_version >= '3.8'
